requirements.txt
# Locked requirements for RAG Evaluation Framework (Python 3.12)
ragas==0.1.21
datasets==2.19.1

# LangChain ecosystem (only if needed by your workflows)
langchain==0.2.11
langchain-openai==0.1.17
langchain-community==0.2.10

python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.5  # For Excel file processing
openai==1.35.13
python-docx==1.1.2  # For Word document processing
PyPDF2==3.0.1  # For PDF processing
pdfplumber==0.11.4  # Alternative PDF processing library
pillow==10.4.0  # For image processing
pytesseract==0.3.10  # For OCR
streamlit==1.36.0  # For manual evaluation interface
matplotlib==3.8.4  # For result visualization
seaborn==0.13.2  # For statistical data visualization